new update

+++ app/Http/Controllers/IdeaController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Idea;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class IdeaController extends Controller
 {
@@ -26,16 +27,16 @@ class IdeaController extends Controller
     /**
      * Store a newly created resource in storage.
      */
-    public function store(Request $request)
+    public function store()
     {
-        //
-        $request->validate([
+        
+        $validated = request()->validate([
             'content' => 'required|min:3|max:240'
         ]);
 
-        $idea = Idea::create([
-            'content' => $request->get('content', '')
-        ]);
+        $validated['user_id'] = auth()->id();
+
+        Idea::create($validated);
         
         return redirect()->route('dashboard')->with('success', 'Idea created successfully!');
 
@@ -46,7 +47,7 @@ class IdeaController extends Controller
      */
     public function show(Idea $idea)
     {
-        return view('idea.show', compact('idea'));
+        return view('idea.show', compact('idea', 'user'));
     }
 
     /**
@@ -54,7 +55,10 @@ class IdeaController extends Controller
      */
     public function edit(Idea $idea)
     {
-        //
+        if(auth()->id() !== $idea->user_id) {
+            abort(404);
+        }
+
         $editing = true;
         return view('idea.show', compact('idea', 'editing'));
     }
@@ -62,15 +66,17 @@ class IdeaController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, Idea $idea)
+    public function update(Idea $idea)
     {
-        //
-        $request->validate([
+        if(auth()->id() !== $idea->user_id) {
+            abort(404);
+        }
+
+        $validated = request()->validate([
             'content' => 'required|min:3|max:240'
         ]);
 
-        $idea->content = $request->get('content', '');
-        $idea->save();
+        $idea->update($validated);
 
         return redirect()->route('idea.show', $idea->id)->with('success', 'Idea updated successfully!');
 
@@ -81,7 +87,11 @@ class IdeaController extends Controller
      */
     public function destroy(Idea $idea)
     {
-       $idea->delete();
-       return redirect()->route('dashboard')->with('success', 'Idea deleted successfully!');
+        if(auth()->id() !== $idea->user_id) {
+            abort(404);
+        }
+
+        $idea->delete();
+        return redirect()->route('dashboard')->with('success', 'Idea deleted successfully!');
     }
 }
